requirements.txt
absl-py==1.2.0
anyio==3.6.1
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
cycler==0.11.0
fastapi==0.82.0
flatbuffers==2.0.7
fonttools==4.37.1
ftfy==6.1.1
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.1
h5py==3.7.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
nltk==3.7
numpy==1.23.2
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pydantic==1.10.2
pymongo==4.2.0
pyparsing==3.0.9
python-datamuse==1.3.2
python-dateutil==2.8.2
pytz==2022.2.1
regex==2022.8.17
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
sniffio==1.3.0
starlette==0.19.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
torch==1.12.1
torchray==1.0.0.2
torchvision==0.13.1
tqdm==4.64.1
typing-extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
git+https://github.com/openai/CLIP.git
